Adding options for multiple decorators
The current implementation only allows to overwrite the adapters. If a user only wants to add a decorator they have to call defaults.apply to build the default adapter based on the given options for different decorators and they have to import the utils module to build it. With this new implementation you just have to pass an array of decorators, which are added in order to the default adapters (based on the passed options)