Fixing ArrayAdapter optgroup children data binding
ArrayAdapter.convertToOptions fails to create item data for optgroup children, not binding it to the select options.  
This results in a bug, where the passed data element to select(data) is not initialized and normalized, thus element is null and id is an integer. Therefore the select function can't find the corresponding <option> while comparing an integer id with string <option>[value]